requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
Django==2.1.3
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
drf-nested-routers==0.91
idna==2.7
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
Markdown==3.0.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pygments==2.2.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
